+++ src/Models/Business/DeliveryMethod.cs
@@ -6,32 +6,32 @@ namespace Cims.WorkflowLib.Models.Business
     public class DeliveryMethod
     {
         /// <summary>
-        /// 
+        /// ID of the delivery method.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
-        /// 
+        /// UID of the delivery method.
         /// </summary>
         public string Uid { get; set; }
 
         /// <summary>
-        /// 
+        /// Name of the delivery method.
         /// </summary>
         public string Name { get; set; }
 
         /// <summary>
-        /// 
+        /// Description of the delivery method.
         /// </summary>
         public string Description { get; set; }
 
         /// <summary>
-        /// 
+        /// Type of the delivery method.
         /// </summary>
         public DeliveryMethodType Type { get; set; }
 
         /// <summary>
-        /// 
+        /// Price of the delivery method.
         /// </summary>
         public decimal Price { get; set; }
     }
